refactor(uri): move LG webhook setup into integration service
